Update file names and format files
