+++ Makefile
@@ -1,7 +1,6 @@
 .PHONY: setup
 setup:
 	pipenv sync --dev
-	cp .env.template .env
 	pipenv run pre-commit install
 	pipenv run pre-commit run --all-files
 	pipenv run pytest
