+++ setup.cfg
@@ -1,0 +1,6 @@
+[flake8]
+# Black style preferred over Flake8 defaults
+ignore =
+    E203,
+    E501,
+    W503,
