+++ src/smart_building_rating_calculator/flexer_enums.py
@@ -1,0 +1,10 @@
+from enum import StrEnum
+
+
+class FlexArchetype(StrEnum):
+    NO_FLEXER = "No Flexer"
+    LOW_TECH_FLEXER = "Low-tech Flexer"
+    GOOD_FLEXER = "Good Flexer"
+    UNTAPPED_FLEXER = "Untapped Flexer"
+    STRONG_FLEXER = "Strong Flexer"
+    GOLD_FLEXER = "Gold Standard Flexer"
